moved api-type-check.spec
